Rename exception package to advice package
